feat: Validation button title refactor. Some files reordered
