+++ Sources/Alloy/Internal/Views/Scan Instructions/Footers/DeniedFooter.swift
