+++ Sources/Alloy/Internal/Views/Scan Instructions/Footers/ScanFooter.swift
