+++ Sources/Alloy/Internal/Views/Scan Instructions/ScanTypes/FrontBackTakePicture.swift
