+++ Sources/Alloy/Internal/Views/Scan Instructions/ScanTypes/PassportTakePicture.swift
