+++ Sources/Alloy/Internal/Views/Scan Instructions/ScanTypes/TakeSelfieView.swift
