+++ Sources/Alloy/Internal/Views/Scan Instructions/ScanTypes/UploadFile.swift
